Fix the comments in PR review * InvalidCourseData now extends RuntimeException * Validation methods are now moved to Validator class
